changed name of the variable form selectStock to stockSelector and moved code from customHook to home page and deleted the customHook file
