+++ src/pages/Home.jsx
@@ -1,10 +1,18 @@
 
 import StockTable from '../components/stockTable';
-import { useStockData } from '../utils/hooks/useStockData';
-
-
+import { useDispatch, useSelector } from "react-redux";
+import { selectStockHeaders, selectTopGainers, selectTopLosers } from "../redux/selectors/stockSelector";
+import { useEffect } from "react";
+import { fetchStockData } from "../redux/reducers/stockReducer";
 export default function Home() {
-const {headers,topGainers,topLosers}=useStockData()
+    const headers = useSelector(selectStockHeaders);
+    const topGainers = useSelector(selectTopGainers);
+    const topLosers = useSelector(selectTopLosers)
+    const dispatch = useDispatch();
+    useEffect(() => {
+        dispatch(fetchStockData())
+    }, [dispatch]);
+
 
     return (
         <div className='w-full min-h-screen p-8 flex gap-4 '>
