+++ src/redux/selectors/stockSelector.js
@@ -1,11 +1,11 @@
 import { createSelector } from "@reduxjs/toolkit";
 
 
-const selectStock = state=>state.stock;
-const selectLoading =createSelector(selectStock,stock=>stock.loading)
-const selectStockData= createSelector(selectStock,stock=>stock.stockData)
+const stockSelector = state=>state.stock;
+const selectLoading =createSelector(stockSelector,stock=>stock.loading)
+const selectStockData= createSelector(stockSelector,stock=>stock.stockData)
 const selectStockHeaders = createSelector([selectStockData], (stockData) => stockData.headers)
-const selectTopGainers=createSelector(selectStock,stock=>stock.topGainers)
-const selectTopLosers=createSelector(selectStock,stock=>stock.topLosers)
+const selectTopGainers=createSelector(stockSelector,stock=>stock.topGainers)
+const selectTopLosers=createSelector(stockSelector,stock=>stock.topLosers)
 
-export { selectStock,selectLoading, selectStockData, selectTopGainers, selectTopLosers, selectStockHeaders };
+export { stockSelector,selectLoading, selectStockData, selectTopGainers, selectTopLosers, selectStockHeaders };
